Generalize even more overlay trigger
