+++ webapp/channels/src/components/admin_console/team_channel_settings/team/details/team_profile.tsx
@@ -16,11 +16,10 @@ import {openModal} from 'actions/views/modals';
 import useGetUsage from 'components/common/hooks/useGetUsage';
 import useGetUsageDeltas from 'components/common/hooks/useGetUsageDeltas';
 import FormattedMarkdownMessage from 'components/formatted_markdown_message';
-import OverlayTrigger from 'components/overlay_trigger';
 import PricingModal from 'components/pricing_modal';
-import {createTooltip} from 'components/tooltip';
 import AdminPanel from 'components/widgets/admin_console/admin_panel';
 import TeamIcon from 'components/widgets/team_icon/team_icon';
+import WithTooltip from 'components/with_tooltip';
 
 import {ModalIdentifiers} from 'utils/constants';
 import {imageURLForTeam} from 'utils/utils';
@@ -34,12 +33,6 @@ type Props = {
     isDisabled?: boolean;
     saveNeeded?: boolean;
 }
-
-const SharedTooltip = createTooltip({
-    id: 'sharedTooltip',
-    title: defineMessage({id: 'workspace_limits.teams_limit_reached.upgrade_to_unarchive', defaultMessage: 'Upgrade to Unarchive'}),
-    hint: defineMessage({id: 'workspace_limits.teams_limit_reached.tool_tip', defaultMessage: 'You\'ve reached the team limit for your current plan. Consider upgrading to unarchive this team or archive your other teams'}),
-});
 
 export function TeamProfile({team, isArchived, onToggleArchive, isDisabled, saveNeeded}: Props) {
     const teamIconUrl = imageURLForTeam(team);
@@ -72,11 +65,11 @@ export function TeamProfile({team, isArchived, onToggleArchive, isDisabled, save
     const button = () => {
         if (restoreDisabled) {
             return (
-                <OverlayTrigger
-                    delay={400}
+                <WithTooltip
+                    id='sharedTooltip'
+                    title={defineMessage({id: 'workspace_limits.teams_limit_reached.upgrade_to_unarchive', defaultMessage: 'Upgrade to Unarchive'})}
+                    hint={defineMessage({id: 'workspace_limits.teams_limit_reached.tool_tip', defaultMessage: 'You\'ve reached the team limit for your current plan. Consider upgrading to unarchive this team or archive your other teams'})}
                     placement='bottom'
-                    disabled={!restoreDisabled}
-                    overlay={<SharedTooltip/>}
                 >
                     {/* OverlayTrigger doesn't play nicely with `disabled` buttons, because the :hover events don't fire. This is a workaround to ensure the popover appears see: https://github.com/react-bootstrap/react-bootstrap/issues/1588*/}
                     <div
@@ -107,8 +100,7 @@ export function TeamProfile({team, isArchived, onToggleArchive, isDisabled, save
                             <FormattedMessage {...archiveBtn}/>
                         </button>
                     </div>
-                </OverlayTrigger>
-
+                </WithTooltip>
             );
         }
         return (
